backend/requirements.txt
django
python-dotenv
google-genai 
requests

